+++ include/host/gemm.h
@@ -1,6 +1,6 @@
 #ifndef CPU_GEMM_H
 #define CPU_GEMM_H
 
-void SgemmWithCPU(int *A, int *B, int *C, int M, int N, int K);
+void SgemmWithCPU(float *A, float *B, float *C, int M, int N, int K);
 
 #endif // CPU_GEMM_H
